#65 flow view page and detailed view page
